starting to rearrange components
