+++ src/app/App.js
@@ -1,9 +1,9 @@
 import './App.css';
-import Garden from '../garden/Garden';
+import Scene from '../scene/Scene';
 import Calendar from '../calendar/Calendar';
-import { PerspectiveCamera, OrbitControls, Sky } from '@react-three/drei';
-import { Canvas } from '@react-three/fiber';
+import Nav from '../nav/Nav';
 import { getHabits } from '../apiCalls';
+import { Routes, Route } from 'react-router-dom';
 // import habits from '../mockData/userHabits';
 import React, { useState, useEffect } from 'react'
 
@@ -30,27 +30,14 @@ function App() {
   }
   
   return (
-    <div style={{ width: '100%', height: '90%' }}>
-      <Canvas className='garden-scene'>
-        <ambientLight intensity={2} />
-        <directionalLight position={[1, 1, 4]} intensity={3} />
-        <Garden habits={userHabits} setError={setError} />
-        <Sky sunPosition={[0.6, 0.1, 0.6]} />
-        <PerspectiveCamera makeDefault position={[0, 5, 15]} />
-        <OrbitControls
-          minDistance={5}
-          maxDistance={25}
-          minPolarAngle={0.5}
-          maxPolarAngle={Math.PI / 1.9}
-          // minAzimuthAngle={-Math.PI / 4}
-          // maxAzimuthAngle={Math.PI / 4}
-          enableZoom={true}
-          enablePan={true}
-          enableRotate={true}
-          />
-      </Canvas>
-          {error && <h2 className="fetch-error">{error.message}</h2>}
-    </div>
+    <>
+      <Nav />
+      <Scene habits={userHabits} setError={setError} />
+      <Routes>
+        <Route path="/calendar" element={<Calendar />} />
+      </Routes>
+      {error && <h2 className="fetch-error">{error.message}</h2>}
+    </>
   );
 }
