+++ src/garden/Garden.js
@@ -1,5 +1,5 @@
 import './Garden.css';
-import { Planter, Dirt, Gnome, Compost, Ground } from '../scene/Scene';
+import { Planter, Dirt, Gnome, Compost, Ground } from '../gardenObjects/GardenObjects';
 import { Stars } from '@react-three/drei';
 import AllFlowers from '../AllFlowers/AllFlowers';
